Refactor Tailscale log collection and output handling
Updated output directory to a fixed path and adjusted log saving method to append to a single log file. Reduced system log lines collected to manage JSON size. Wazuh should be able to read the JSON files now that they are on one line and appended to a single file name